+++ src/Annotation/RefSource.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace Kassko\DataMapper\Annotation;
+
+/**
+* @Annotation
+* @Target({"PROPERTY"})
+*
+* @author kko
+*/
+final class RefSource
+{
+	/**
+     * Id of a source (data source, provider)
+     * @var string
+     */
+    public $ref;
+}