+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+`0.12.1.0-alpha - 2015/03/25 - New version`:
+* `New`: Add DataSourcesStore and ProvidersStore to centralize all sources.
+* `New`: Allows to link a property to source from the store to reduce the mapping code.
+* `New`: Allows to chaine sources as fallback sources if some sources do not respond properly.
+* `Break`: Use stable version of class-resolver component (not an alpha version)
+
 `0.12.0.0-alpha - 2015/03/24 - Break version`:
 * `Break`: To decide if a field should be managed, in your mapping configuration use the keys Include/Exclude/Object.fieldExclusionPolicy. The key Field don't do the job anymore.
 * `Break`: To send the curent object as parameter of a data source method use ##this instead of @this. To send the value of a property, use #some_property instead of some_property.